+++ MacForge/MacForge/Plugins/MF_PluginManager.h
@@ -6,6 +6,11 @@
 //  Copyright © 2016 Wolfgang Baird. All rights reserved.
 //
 
+@import AppKit;
+#import "FConvenience.h"
+#import "MF_Plugin.h"
+#import "MF_defines.h"
+
 @interface MF_PluginManager : NSObject <NSURLSessionDataDelegate, NSURLSessionDelegate, NSURLSessionTaskDelegate, NSUserNotificationCenterDelegate> {
     NSMutableArray      *pluginsArray;
     NSMutableDictionary *installedPluginDICT;